Two Stage Regression Imputation Assignment
For this assignment I was tasked with performing LASSO regression on the variable 'Income' in the Income Data dataset to determine the likelihood of its missing values being 0 or 1, and subsequently performing imputation on the estimated non-zero missing values. To conclude, I was tasked with using linear regression to estimate Income values based on other variables.